Client: Don't create cursor surface until we can set its role
We can't set the cursor role until we have gotten an enter event, so
wait with creating the surface until we want to set the role for the
first time.

Makes testing easier since we don't have to deal with role-less
surfaces that may or may not be the cursor surface.

Change-Id: I99803d9b0b5933a271b6b1d56f74d4ff79813e40
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: David Edmundson <davidedmundson@kde.org>