Revert "Client: Full implementation for frame callbacks"
This caused regressions because QtQuick depends on swapBuffers for throttling
animations. We probably need to emulate a blocking swapBuffers and continue
after a timeout, but until we have a patch for this, revert this to avoid
releasing a regression.

This brings back the bug with a frozen event loop when a surface is waiting for
a frame callback, but this is preferable to a regression.

This reverts commit 58d65cf60ab6f8b33f691d91e8145baf7837fb6e.

Fixes: QTBUG-72578
Change-Id: If6435a947aae5e9fd775404649a392bfafe9130a
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>