Client: Don't leak wl_data_offers
[ChangeLog][QPA plugin] Fixed a leak of wl_data_offers.

Data offers would previously leak when a surface with keyboard focus was
destroyed... or if it was hidden... or if it changed type, and so on.

Make keyboard focus follow a wl_surface instead of a QWaylandWindow.

This also fixes a couple of other minor issues, such as the mRepeatTimer not
stopping when a wl_surface was destroyed.

Ideally, we would have a QWaylandSurface class separate from the
QWaylandWindow, but that's out of scope for this fix.

Fixes: QTBUG-73825
Change-Id: I56e502512c3959e3fcdb63744adc4a7698e3d53d
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>