Avoid creating decoration whilst Xdg Decoration is pending configure
Currently even when we are on on a compositor with SSDs we always create
a decoration instance only to discard it.

Not only is this somewhat wasteful, it creates sizing problems as the
frame geometry changes, leading to constantly expanding windows on kwin.

This patch assumes that if we have a decoration manager we should have
no decoration until it is configured to do so.

This is safe because we request a mode in the constructor, a compositor
must reply with a configure event and we shouldn't be showing any
buffers until the first configure event is received.

Behavior without a decoration manager is unchanged

Change-Id: I72b2cf4423fe6461ba405612262f76cefe4d6201
Reviewed-by: Johan Helsing <johan.helsing@qt.io>