QWaylandCursor: replace a static QMap with a C array
The content of the QMultiMap cursorNameMaps is static, so using a
dynamic container is overkill.

Replace the map by a (sorted) C array of struct {Key, Value}, and use
std::equal_range to find matching cursors. This also avoids the
creation of a QList on each call (due to the use of QMap:values()
instead of QMap::equal_range), as well as the creation of QByteArrays,
which the underlying wayland C API doesn't know about, anyway.

The entries were already sorted, so the initializer_list stays the
same. Add an assert that the list is ordered.

Since this is library code, don't use the attractive struct {
WaylandCursor, const char *name }, as that causes relocations that
delay the loading of the shared object, and make sharing of the data
between processes impossible. This wastes some memory, due to all the
NUL padding up to the common maximum, and the next alignment step, but
can be optimized later when someone writes a C++-only version of
Q_STRINGTABLE.

Even so, this saves 10.5KiB in text size on optimized AMD64 Linux GCC
9.1 builds.

Change-Id: I19fa1742e2fcaf9de74af1e5908e796b0588d507
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>