Convert most config tests to configure.json inline format
And remove the now unused entries in config.tests. Now only the wayland-scanner
test remains, as that uses qmake features that do not easily translate to the
inline format.

This makes it much easier to port QtWayland to cmake.

This also removes some config.tests that weren't used.

Task-number: QTBUG-78177
Change-Id: Ia21a0025f9fac45b9f4815ee0f196e9e5759cdc6
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>