Client: Make frame back callback timeout configurable
To avoid rendering when windows are minimized, we try to detect
this case by checking if 100 ms passes after an update before
any frame callback arrives. Wayland expects to drive the rendering
through frame callbacks, while Qt expects to be informed when a
window is no longer visible and will run a tight, vsynced render loop
until it gets the appropriate event. This mismatch causes issues,
and the timeout is an attempt to avoid actively rendering to a hidden
surface by detecting the missing callbacks.

It causes problems on embedded, though, when a device is so
busy that the timeout happens even when the window is visible.

So we introduce a way to configure the timeout. Either increase
its duration if you can guarantee a certain minimum time between
events even with high load, or set it to 0 to disable it completely.
(at the expense of drawing to invisible windows).

Note that this is required for fixing a critical memory leak that
happens when using Wayland on embedded.

This was mistakenly pushed to 5.12 branch, so it will not be
automatically merged upstream. Therefore we have to
cherry-pick it.

[ChangeLog][Client] Added support for QT_WAYLAND_FRAME_CALLBACK_TIMEOUT
environment variable, which can be used to disable or change the
internal frame callback timeout. If you see windows that stop rendering
or minimize on heavy load, then try setting the variable to 0.

Task-number: QTBUG-82914
Reviewed-by: David Edmundson <davidedmundson@kde.org>
(cherry picked from commit 0a0e0eb8dca699858435dec5194b0b8b6ebd3cf8)
Change-Id: I2aeeecb0fab8f7be9b838e477c88eae22b322d75
Reviewed-by: Janne Koskinen <janne.p.koskinen@qt.io>