Client: Send subsurface expose event when toplevel is configured
If a subsurface is set to be visible on the cilent side before the top
level is configured it will do not create an exposeEvent and map a
buffer as we fail the check in isExposed() where we check the parent.
This is correct behavior.

However, when the toplevel receives an applyConfigure from the shell
client we need subsurfaces to update accordingly.

This fixes a race where subsurfaces are not shown with slow compositors.

Change-Id: Icd156e7655d5b25535acc4d2fe77c31e19ebfa32
Pick-to: 5.15
Fixes: QTBUG-86176
Reviewed-by: David Edmundson <davidedmundson@kde.org>