CMake: Make sure test blacklists are found
Some of the qtwayland tests are failing on qemu configs and are thus
blacklisted. When the tests were built with CMake, the blacklist file was
not found.

This is due to the custom QCOMPOSITOR_TEST_MAIN function which does not
provide the location of the test source folder, and
thus the QTestPrivate::parseBlackList -> QTest::qFindTestData call
fails to find black list file.

Bring the custom test main function in line with what the other QTest
main functions do, which is call QTEST_SET_MAIN_SOURCE_PATH.

Note in qmake builds this was not necessary, because the way the
source and build folders are computed for tests is slightly different
from the CMake way. I don't recall the exact details, but
basically CMake builds need this extra information that is set by the
QTEST_SET_MAIN_SOURCE_PATH call.

Task-number: QTBUG-83968
Change-Id: Ia4742488097a65f18594881c89af48c9d621e6e1
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>