Re-visit the xkbcommon dependency
QtGui doesn't expose xkbcommon as public dependency anymore.
QtGuiPrivate, however does now, but only in the CMake build.

Fix the qmake build by adding an explicit xkbcommon dependency to
client.pro, like it's done for the compositor.

Commented out the explicit xkbcommon dependencies in the CMakeLists.txt
files of client and compositor.

Mark the XKB packages as optional, such that qtwayland can actually be
built if the feature xkbcommon is OFF and no libxkbcommon-dev package is
installed.

Change-Id: Ia1c7a8333dc8a93b4924e03f7f2284a709a9d65b
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>