Decorations: add new argument to margins() to allow exclude shadows
We can already have shadows to be part of decoration margins, but we
need a way to also get size of shadow margins itself, because they
should be ignored for the purposes of aligning, placing and constraining
windows.

This change breaks API/ABI compatibility of QWaylandAbstractDecoration
plugin, however, as stated in the header file, it is a subject of change
and its API is not meant to be stable. Alternatively we can make this a
separate method.

Change-Id: Ib2c9982e16c4acb13a83757e6371022c0a89b74e
Reviewed-by: David Edmundson <davidedmundson@kde.org>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>