QWaylandClientExtension: Allow specifying destructor for wayland objects
While the template takes care of creating proxies automatically,
destroying them is harder since an interface can have 0 to multiple
destructors. However in the most common case there is only one or always
calling one is sufficient.
An additional template parameter is introduced that allows user code to
specify a callable taking a pointer to the scanner generated class that
should be called when the wayland object is to be be destroyed. This is
done when the global is removed or upon destruction of the C++ object
itself. The clientextension test is changed how it can be used.
Since it works via a non-type template parameter a pointer to a (member)
function can be passed or when compiled in c++20 mode a lambda or
for example a function object.
This new functionality is opt-in and the default behavior is unchanged.
The default value is nullptr used as a tag to do not enable the new
behavior.

Change-Id: I8043f7106fec0cd6f2a79682e5872cfa8da3d11b
Reviewed-by: David Edmundson <davidedmundson@kde.org>