Client: Stop requesting activation on every focus object change
Currently, windows with xdg-toplevel surface role request window activation
every time the focus object changes. It conflicts with the compositor's
activation and stacking order policies.

For example, when a window is minimized, the compositor will start an
animation and move focus to the next window. If that window activates
itself, it can be raised in the stack and start covering the minimized
window, thus making it look as if no animation is played at all.

This activation logic was introduced as part of qt-shell activation
model in 260b55e2b93d65d847fef44ffa06bedaee048507. On the other hand,
with the introduction of the xdg-activation-v1 protocol, there is a
better way to pass input focus, so drop the qt-shell hack in favor of
that.

Fixes: QTBUG-111377
Pick-to: 6.5
Change-Id: I8ed19f1be17f49fa9748b1cb5fdd0070791d2ecd
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>