Client: Track parent-popup relationship in QWaylandWindow
This reduces the amount of boilerplate code that goes in shell
integration plugins providing popups.

The main motivation behind this change though is to ensure that
QWaylandWindow::addChildPopup() gets called outside the
QWaylandShellSurface constructor so one could use the popup's shell
surface object to customize parent-child relationship.

    mShellSurface = mShellIntegration->createShellSurface(this);

when this code executes, addChildPopup() will be called before the
return value of createShellSurface() is assigned to mShellSurface.

Change-Id: I9ccfb21f46febb451bdd7b4aa7851a99f3a03655
Reviewed-by: David Edmundson <davidedmundson@kde.org>