Client: Remove check for BypassWindowManagerHint when setting viewport
This check was put in initially to leave a code path for clients to
opt-out of Qt controlled viewport management. In practice, many X11
applications have this flag set for other purposes and this does more
harm than good.

Given this was never a documented feature, and it won't have been used
in practice given fractional scaling is not heavily deployed outside a
few linux desktops it is safe to remove.

The highDpiScaleFactorRoundingPolicy remains a way to opt-out.

Pick-to: 6.5
Change-Id: I2d111a76695dec27e9ee1a2a233252b70379a329
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>