Don't floor device scale in `QWaylandShmBuffer::imageInsideMargins`
The offset of the drawing target returned from
QWaylandShmBackingStore::contentSurface() is incorrectly calculated due
to `mImage.devicePixelRatio()` being casted to `int` inside
`imageInsideMargins`, which causes the scale to be rounded down in case
of fractional scaled displays. This causes the window contents to be
incorrectly positioned on CSD-requiring compositors (e.g. GNOME),
eventually resulting in graphics corruption due to mismatched damage and
opaque regions.

Fix this by removing the incorrect `int` cast.

Pick-to: 6.5
Change-Id: I436d24522d23888b641cb21403f773e8ea113b8d
Reviewed-by: David Edmundson <davidedmundson@kde.org>