client: Guard against client destruction in gestures
Client code can delete a window at any point. Any gesture events,
including a "begin" event can be already in flight before the compositor
is aware of the window destruction.

It's up to the client code to handle either the focus not being
initially found or the QPointer resetting and no-op.

Task-number: QTBUG-113145
Pick-to: 6.6 6.5
Change-Id: Ie2c01799bd38c6f295159876a1bcd018abe60188
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>