client: Simplify toplevel exposure tracking
m_exposeRegion was originally there to track the expose updates
requested from the client in situations where the shell integration is
filtering those events, i.e when we're waiting for the first configure
event.

The fix at e4c0f02b54201842f20abd5c903c627a4d6ed2ff explicitly sends an
expose event to the whole window when we get the actual first configure
event, at which point caching the expose region at the time of the
handleExpose check is unnecessary.

Motivation for cleanup is a pending change to support
xdg_toplevel.suspended where we need to unexpose at a shell level.

Change-Id: I56c3bab91453956046c1d7f92a4395c951c48fb3
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>