qtwaylandscanner: Send null for a null QString with allow-null
`toUtf8().constData()` on a null `QString` results in an empty
string sent over the wire, which is distinct from `null` when
the request has "allow-null" specified.

Notably `wl_data_offer.accept` uses a null mime_type for
"not accepted".

[ChangeLog][Important Behavior Changes][qtwaylandscanner] String
arguments in a request will now send null rather than empty string
for a null QString if the argument is declared with "allow-null".

Change-Id: I4b78246e4da7b60680d7797fd6309755f44d2bb6
Reviewed-by: David Edmundson <davidedmundson@kde.org>