client: Workaround Gnome's incorrect xdg-geometry sizes
Mutter, depending on the scaling mode, will sometimes send the xdg
geometry size in device pixels rather than logical pixels.

It is reported as https://gitlab.gnome.org/GNOME/mutter/-/issues/2631
but it is closed upstream so we cannot rely on them fixing it.

This patch introduces the same workaround that GTK ships to work around
this behavior. If the scale is set, and the xdg geometry matches the
output geometry, then we can assume we don't have logical pixels for the
xdg geometry and should calculate it manually.

Fixes: QTBUG-122197
Change-Id: I7f640bac619be6eb07c84bb6913334a6356524a2
Pick-to: 6.8 6.7 6.5
Reviewed-by: Aleix Pol Gonzalez <aleixpol@kde.org>