client: Cache exposed state and unify sending events
Right now we are sending an expose event on every configure event
because exposure may have changed. This causes an overhead.

The mechanism for sending expose events is also scattered, there are a
lot of properties that affect exposure, having everyone track the full
state gets messy.

This is cleanup work to then be able to land viewport clipping and
synchronous event delivery.

Change-Id: I192b908cdb6f6adda7d0e33f83961861e09504e8
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>