Client: Use std::unique_ptr to manage window decoration
There was a potential crash in the QWaylandWindow destructor. First it
deletes the window decoration (but does not unset the pointer!) and then
deletes the surface via reset(). This eventually triggers
QWaylandDisplay::handleKeyboardFocusChanged which will call
decoration->update().
Using std::unique_ptr makes it easier to manage and prevents such errors.

Change-Id: I52edcdab0d25c35dc656581fbb5c1c9b5ca481c4
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>