client: Drop outdated QT_WAYLAND_RESIZE_AFTER_SWAP
QT_WAYLAND_RESIZE_AFTER_SWAP was added
in15f82d0bc40eae635ed1377ae39253a8f515ea4a in 2014 to provide an
opt-in-path to support drivers that did not apply an eglWindow resize
until after the swap occurred.

This is now the default in the EGL specification and fully supported in
QtWayland. The environment variable does nothing.

Change-Id: I8114d11801db5ba6863f48e2d44cca984bb58de7
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>